refactor: organise components/ folder
